transport commit ...made so changes

+++ ViewerWidget.cpp
@@ -691,8 +691,8 @@ void ViewerWidget::drawObject(Object_H_edge object, Camera camera, ProjectionPla
 Object_H_edge ViewerWidget::perspectiveCoordSystemTransformation(Object_H_edge object) {
 	Object_H_edge transformedObject = object;
 	for (Vertex* vertex : transformedObject.vertices) {
-		vertex->x = img->width() / 2  - *vertex * projectionPlane.basisVectorV - camera.position.x;	//overload vector dot product 
-		vertex->y = img->height() / 2  - *vertex * projectionPlane.basisVectorU - camera.position.y;
+		vertex->x = static_cast<double>(img->width() / 2)  - (*vertex) * projectionPlane.basisVectorV - camera.position.x;	//overload vector dot product 
+		vertex->y = static_cast<double>(img->height() / 2)  - (*vertex) * projectionPlane.basisVectorU - camera.position.y;
 		vertex->z = *vertex * projectionPlane.basisVectorN - camera.position.z;
 	}
 	return transformedObject;
